# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KESHAV",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/locations": {
      "get": {
        "tags": [
          "locations"
        ],
        "summary": "List Locations",
        "operationId": "list_locations_api_v1_locations_get",
        "parameters": [
          {
            "name": "location_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LocationType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/locations/{location_id}": {
      "get": {
        "tags": [
          "locations"
        ],
        "summary": "Get Location",
        "operationId": "get_location_api_v1_locations__location_id__get",
        "parameters": [
          {
            "name": "location_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Location Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Location"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/weather": {
      "get": {
        "tags": [
          "weather"
        ],
        "summary": "Get Weather",
        "operationId": "get_weather_api_v1_weather_get",
        "parameters": [
          {
            "name": "location_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location Id"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO-8601 start (UTC)",
              "title": "Start"
            },
            "description": "ISO-8601 start (UTC)"
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO-8601 end (UTC)",
              "title": "End"
            },
            "description": "ISO-8601 end (UTC)"
          },
          {
            "name": "seed",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 2026,
              "title": "Seed"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherObservationList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/forecast": {
      "get": {
        "tags": [
          "forecast"
        ],
        "summary": "Get Forecast",
        "operationId": "get_forecast_api_v1_forecast_get",
        "parameters": [
          {
            "name": "location_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location Id"
            }
          },
          {
            "name": "issued_at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO-8601 issue time (UTC)",
              "title": "Issued At"
            },
            "description": "ISO-8601 issue time (UTC)"
          },
          {
            "name": "horizon_hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 240,
              "minimum": 1,
              "default": 120,
              "title": "Horizon Hours"
            }
          },
          {
            "name": "seed",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 2026,
              "title": "Seed"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherForecastList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ingestion/logs": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Ingestion Logs",
        "operationId": "ingestion_logs_api_v1_ingestion_logs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IngestionRunDetail"
                  },
                  "title": "Response Ingestion Logs Api V1 Ingestion Logs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IngestionRunDetail": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "ingestion_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ingestion Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "status": {
            "$ref": "#/components/schemas/IngestionStatus",
            "default": "STARTED"
          },
          "records_received": {
            "type": "integer",
            "title": "Records Received",
            "default": 0
          },
          "records_accepted": {
            "type": "integer",
            "title": "Records Accepted",
            "default": 0
          },
          "records_rejected": {
            "type": "integer",
            "title": "Records Rejected",
            "default": 0
          },
          "records_flagged": {
            "type": "integer",
            "title": "Records Flagged",
            "default": 0
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "quality_summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Summary"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "fallback_used": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Used"
          },
          "fallback_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Source"
          }
        },
        "type": "object",
        "required": [
          "source",
          "provider",
          "domain",
          "started_at"
        ],
        "title": "IngestionRunDetail",
        "description": "One ingestion run including Step 2 latency/fallback metadata."
      },
      "IngestionStatus": {
        "type": "string",
        "enum": [
          "STARTED",
          "SUCCESS",
          "PARTIAL",
          "FAILED"
        ],
        "title": "IngestionStatus"
      },
...
      "Location": {
        "properties": {
          "location_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 3,
            "title": "Location Id"
          },
          "parent_location_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Location Id"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "type": {
            "$ref": "#/components/schemas/LocationType"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "spatial": {
            "$ref": "#/components/schemas/SpatialReference"
          },
          "population": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Population"
          },
          "area_sq_km": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Area Sq Km"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "Where the location came from",
            "default": ""
          },
          "provenance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Provenance"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "last_updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Updated At"
          }
        },
        "type": "object",
        "required": [
          "location_id",
          "name",
          "type"
        ],
        "title": "Location",
        "description": "Canonical location record exposed via the API."
      },
      "LocationList": {
        "properties": {
          "locations": {
            "items": {
              "$ref": "#/components/schemas/Location"
            },
            "type": "array",
            "title": "Locations"
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "locations"
        ],
        "title": "LocationList"
      },
      "LocationType": {
        "type": "string",
        "enum": [
          "country",
          "state",
          "district",
          "city",
          "ward",
          "zone",
          "neighborhood",
          "point"
        ],
        "title": "LocationType"
      },
...
      "Provenance": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source",
            "description": "Source system, e.g. 'imd', 'mock_weather'"
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "description": "Provider name inside the source"
          },
          "adapter": {
            "type": "string",
            "title": "Adapter",
            "description": "Adapter that produced the canonical record",
            "default": ""
          },
          "source_timestamp": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Timestamp"
          },
          "ingestion_timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Ingestion Timestamp",
            "description": "When KESHAV ingested this record"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version",
            "description": "Provider/model version"
          },
          "quality_status": {
            "$ref": "#/components/schemas/QualityFlag",
            "default": "VALID"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "source",
          "provider"
        ],
        "title": "Provenance",
        "description": "Traceability for every data record.\n\nHelps answer: \"Where did this value come from?\""
      },
...
      "QualityFlag": {
        "type": "string",
        "enum": [
          "VALID",
          "SUSPECT",
          "MISSING",
          "STALE",
          "INVALID",
          "IMPUTED"
        ],
        "title": "QualityFlag",
        "description": "Structured data-quality flags attached to every observation.\n\nMISSING means the value was absent (never converted to zero).\nIMPUTED means a later pipeline stage filled the gap \u2014 never done silently."
      },
...
      "SpatialReference": {
        "properties": {
          "ref_type": {
            "$ref": "#/components/schemas/SpatialReferenceType",
            "default": "point"
          },
          "resolution_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Value",
            "description": "Size of one cell/resolution unit"
          },
          "resolution_unit": {
            "$ref": "#/components/schemas/SpatialResolutionUnit",
            "default": "degrees"
          },
          "geometry_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geometry Reference",
            "description": "External geometry id/URL (GIS integration later)"
          },
          "crs": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crs",
            "description": "Coordinate reference system"
          }
        },
        "type": "object",
        "title": "SpatialReference",
        "description": "How a record is located in space.\n\nNo scientific interpolation is performed in Step 1; this abstraction simply\nrecords the geometry type and native resolution so later engines can align."
      },
      "SpatialReferenceType": {
        "type": "string",
        "enum": [
          "point",
          "raster_cell",
          "polygon",
          "grid",
          "practice"
        ],
        "title": "SpatialReferenceType"
      },
      "SpatialResolutionUnit": {
        "type": "string",
        "enum": [
          "degrees",
          "km",
          "m"
        ],
        "title": "SpatialResolutionUnit"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WeatherForecastList": {
        "properties": {
          "forecasts": {
            "items": {
              "$ref": "#/components/schemas/WeatherForecast"
            },
            "type": "array",
            "title": "Forecasts"
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "forecasts"
        ],
        "title": "WeatherForecastList"
      },
...
      "WeatherObservationList": {
        "properties": {
          "observations": {
            "items": {
              "$ref": "#/components/schemas/WeatherObservation"
            },
            "type": "array",
            "title": "Observations"
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "observations"
        ],
        "title": "WeatherObservationList"
      }
    }
  }
}